+++ rudehash.ps1
@@ -278,6 +278,7 @@ $Pools =
 			"neoscrypt" = @{ Server = "neoscrypt.mine.zpool.ca"; Port = 4233; Modifier = 1000000 }
 			"nist5" = @{ Server = "nist5.mine.zpool.ca"; Port = 3833; Modifier = 1000000 }
 			"phi" = @{ Server = "phi.mine.zpool.ca"; Port = 8333; Modifier = 1000000 }
+			"phi2" = @{ Server = "phi2.mine.zpool.ca"; Port = 8332; Modifier = 1000000 }
 			"skein" = @{ Server = "skein.mine.zpool.ca"; Port = 4933; Modifier = 1000000 }
 			"tribus" = @{ Server = "tribus.mine.zpool.ca"; Port = 8533; Modifier = 1000000 }
 			# "polytimos" = @{ Server = "polytimos.mine.zpool.ca"; Port = 8463; Modifier = 1000000 }
@@ -304,7 +305,7 @@ $Coins =
 	"hsr" = @{ Name = "Hshare"; Algo = "hsr" }
 	"ifx" = @{ Name = "Infinex"; Algo = "lyra2z" }
 	"kreds" = @{ Name = "Kreds"; Algo = "lyra2v2" }
-	"lux" = @{ Name = "LUXCoin"; Algo = "phi"; WtmId = 212 }
+	"lux" = @{ Name = "LUXCoin"; Algo = "phi2"; WtmId = 212 }
 	"mona" = @{ Name = "Monacoin"; Algo = "lyra2v2"; WtmId = 148 }
 	"pgn" = @{ Name = "Pigeoncoin"; Algo = "x16s" }
 	"poly" = @{ Name = "Polytimos"; Algo = "polytimos" }
@@ -327,7 +328,7 @@ $Miners =
 	"ccminer-phi" = @{ Url = "https://github.com/216k155/ccminer-phi-anxmod/releases/download/ccminer%2Fphi-1.0/ccminer-phi-1.0.zip"; ArchiveFile = "ccminer-phi.zip"; ExeFile = "ccminer.exe"; FilesInRoot = $false; Algos = @("phi"); Api = $true; Version = "1.0" }
 	"ccminer-polytimos" = @{ Url = "https://github.com/punxsutawneyphil/ccminer/releases/download/polytimosv2/ccminer-polytimos_v2.zip"; ArchiveFile = "ccminer-polytimos.zip"; ExeFile = "ccminer.exe"; FilesInRoot = $true; Algos = @("polytimos"); Api = $true }
 	"ccminer-rvn" = @{ Url = "https://github.com/MSFTserver/ccminer/releases/download/2.2.5-rvn/ccminer-x64-2.2.5-rvn-cuda9.7z"; ArchiveFile = "ccminer-rvn.7z"; ExeFile = "ccminer-x64.exe"; FilesInRoot = $true; Algos = @("x16r"); Api = $true; Version = "2.2.5" }
-	"ccminer-tpruvot" = @{ Url = "https://github.com/tpruvot/ccminer/releases/download/2.2.4-tpruvot/ccminer-x64-2.2.4-cuda9.7z"; ArchiveFile = "ccminer-tpruvot.7z"; ExeFile = "ccminer-x64.exe"; FilesInRoot = $true; Algos = @("bitcore", "equihash", "hsr", "keccakc", "lyra2v2", "lyra2z", "neoscrypt", "nist5", "phi", "polytimos", "skein", "tribus"); Api = $true; Version = "2.2.4" }
+	"ccminer-tpruvot" = @{ Url = "https://github.com/tpruvot/ccminer/releases/download/2.3-tpruvot/ccminer-2.3-cuda9.7z"; ArchiveFile = "ccminer-tpruvot.7z"; ExeFile = "ccminer-x64.exe"; FilesInRoot = $true; Algos = @("bitcore", "equihash", "hsr", "keccakc", "lyra2v2", "lyra2z", "neoscrypt", "nist5", "phi", "phi2", "polytimos", "skein", "tribus"); Api = $true; Version = "2.3" }
 	"ccminer-xevan" = @{ Url = "https://github.com/krnlx/ccminer-xevan/releases/download/0.1/ccminer.exe"; ArchiveFile = "ccminer-xevan.exe"; ExeFile = "ccminer-xevan.exe"; FilesInRoot = $true; Algos = @("xevan"); Api = $true }
 	"dstm" = @{ Url = "https://github.com/nemosminer/DSTM-equihash-miner/releases/download/DSTM-0.6/zm_0.6_win.zip"; ArchiveFile = "dstm.zip"; ExeFile = "zm.exe"; FilesInRoot = $false; Algos = @("equihash"); Api = $true; Version = "0.6" }
 	"ethminer" = @{ Url = "https://github.com/ethereum-mining/ethminer/releases/download/v0.14.0rc2/ethminer-0.14.0rc2-Windows.zip"; ArchiveFile = "ethminer.zip"; ExeFile = "ethminer.exe"; FilesInRoot = $false; Algos = @("ethash"); Api = $true; Version = "0.14.0rc2" }
@@ -373,6 +374,7 @@ $WtmModifiers =
 	"neoscrypt" = 1000
 	"nist5" = 1000000
 	"phi" = 1000000
+	"phi2" = 1000000
 	"skein" = 1000000000
 	"tribus" = 1000000
 	"xevan" = 1000000
@@ -391,6 +393,7 @@ $AlgoNames =
 	"neoscrypt" = "NeoScrypt"
 	"nist5" = "Nist5"
 	"phi" = "PHI1612"
+	"phi2" = "PHI2"
 	"polytimos" = "Polytimos"
 	"skein" = "Skein"
 	"tribus" = "Tribus"