+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## RudeHash 9.0 (TBD)
 
-* New algos: Skein
+* New algos: PHI2, Skein
+* Updates: ccminer-tpruvot 2.3, PowerShell 6.1.0-preview4
 * Change DNS resolver to 1.1.1.1
 * Fix Tribus missing from Config Editor